+++ gtsam_unstable/navigation/EqVIOSymmetry.h
@@ -22,6 +22,8 @@
 #include <gtsam_unstable/navigation/EqVIOState.h>
 #include <gtsam_unstable/dllexport.h>
 
+#include <functional>
+
 namespace gtsam {
 namespace eqvio {
 
@@ -50,6 +52,35 @@ GTSAM_UNSTABLE_EXPORT VIOState integrateSystemFunction(
 /// Generate ideal camera measurements from state.
 GTSAM_UNSTABLE_EXPORT VisionMeasurement measureSystemState(
     const VIOState& state, const std::shared_ptr<const VIOCameraModel>& camera);
+
+/// InvDepth EqF coordinate suite and associated matrices/lifts.
+struct GTSAM_UNSTABLE_EXPORT EqFCoordinateSuite {
+  std::function<Vector(const VIOState&, const VIOState&)> stateChart;
+  std::function<VIOState(const Vector&, const VIOState&)> stateChartInv;
+  std::function<Matrix(const VIOGroup&, const VIOState&, const IMUInput&)>
+      stateMatrixA;
+  std::function<Matrix(const VIOGroup&, const VIOState&)> inputMatrixB;
+  std::function<Matrix23(const Point3&, const SOT3&,
+                         const std::shared_ptr<const VIOCameraModel>&,
+                         const Point2&)>
+      outputMatrixCiStar;
+  std::function<Vector(const Vector&, const VIOState&)> liftInnovation;
+  std::function<VIOGroup(const Vector&, const VIOState&)> liftInnovationDiscrete;
+
+  Matrix outputMatrixC(const VIOState& xi0, const VIOGroup& X,
+                       const VisionMeasurement& y,
+                       const std::shared_ptr<const VIOCameraModel>& camera,
+                       bool useEquivariance = true) const;
+
+  Matrix stateMatrixADiscrete(const VIOGroup& X, const VIOState& xi0,
+                              const IMUInput& imuVel, double dt) const;
+
+  Matrix23 outputMatrixCi(const Point3& q0, const SOT3& QHat,
+                          const std::shared_ptr<const VIOCameraModel>& camera)
+      const;
+};
+
+extern const GTSAM_UNSTABLE_EXPORT EqFCoordinateSuite EqFCoordinateSuite_invdepth;
 
 /// Right action phi(xi, X) = stateGroupAction(X, xi).
 struct GTSAM_UNSTABLE_EXPORT VIOSymmetry
